requirements.txt
fastapi>=0.88.0
uvicorn>=0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
qrcode[svg]==7.4.2
Pillow>=10.0.0
reportlab==4.0.7
python-magic-bin==0.4.14
anyio>=4.5.0
starlette>=0.41.3 